Refactor cli package layout
